feat: adding middleware to handle thrown errors

+++ src/modules/functionaries/service/DeleteFunctionaryService.ts
@@ -1,6 +1,7 @@
-import { FunctionaryRepository } from '@modules/functionaries/infra/typeorm/repositories/FunctionaryRepository';
-
 import { getCustomRepository } from 'typeorm';
+
+import { FunctionaryRepository } from '@modules/functionaries/infra/typeorm/repositories/FunctionaryRepository';
+import { AppError } from '@shared/errors/AppError';
 
 interface IRequestDeleteFunctionaryService {
   cpf: string;
@@ -15,7 +16,14 @@ class DeleteFunctionaryService {
   public async execute({
     cpf,
   }: IRequestDeleteFunctionaryService): Promise<Boolean> {
-    return !!(await this.ormFunctionaryRepository.deleteFunctionary(cpf));
+    if (!cpf) throw new AppError('cpf must be provided');
+
+    const deleteWithSucess =
+      !!(await this.ormFunctionaryRepository.deleteFunctionary(cpf));
+
+    if (!deleteWithSucess) throw new AppError('failed to delete functionary');
+
+    return deleteWithSucess;
   }
 }
